+++ pkg/dbcrypt/dbcrypt.go
@@ -6,10 +6,9 @@ import (
 	"crypto/rand"
 	"encoding/hex"
 	"fmt"
+	"github.com/greenbone/opensight-golang-libraries/pkg/dbcrypt/config"
 	"io"
 	"reflect"
-
-	"github.com/greenbone/opensight-golang-libraries/pkg/postgres/dbcrypt/config"
 )
 
 type DBCrypt[T any] struct {
